backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
whitenoise>=6.6.0
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
gunicorn>=21.2.0
Pillow>=10.0.0
pyotp>=2.9.0
qrcode>=7.4.2
channels>=4.0.0,<5.0.0
daphne>=4.0.0,<5.0.0
celery>=5.3.0,<6.0.0
redis>=5.0.0
django-celery-beat>=2.6.0
psycopg2-binary>=2.9.0

